Cleanup int types passed to SST as Dims

+++ source/adios2/toolkit/sst/cp/ffs_marshal.c
@@ -638,8 +638,8 @@ void SstReaderInitFFSCallback(SstStream Stream, void *Reader,
 
 extern void SstFFSGetDeferred(SstStream Stream, void *Variable,
                               const char *Name, size_t DimCount,
-                              const unsigned long *Start,
-                              const unsigned long *Count, void *Data)
+                              const size_t *Start, const size_t *Count,
+                              void *Data)
 {
     struct FFSReaderMarshalBase *Info = Stream->ReaderMarshalData;
     int GetFromWriter = 0;
@@ -1315,9 +1315,8 @@ static void FFSBitfieldSet(struct FFSMetadataInfoStruct *MBase, int Bit)
 
 extern void SstFFSMarshal(SstStream Stream, void *Variable, const char *Name,
                           const char *Type, size_t ElemSize, size_t DimCount,
-                          const unsigned long *Shape,
-                          const unsigned long *Count,
-                          const unsigned long *Offsets, const void *data)
+                          const size_t *Shape, const size_t *Count,
+                          const size_t *Offsets, const void *data)
 {
 
     struct FFSMetadataInfoStruct *MBase;
